Create README.md
The main goal of the project was to understand non-invasive methods for improving drug delivery to the brain. This was created as a part of a year-long academic research initiative. Emphasis was placed on mechanism comprehension, safety considerations, and translational feasibility than experimental implementation. In order to assess focused ultrasound and microbubble-assisted techniques in relation to CNS disorders and brain tumors, the study combined the data from peer reviewed literature.